+++ qt_ui/models.py
@@ -547,6 +547,7 @@ class GameModel:
         else:
             self.ato_model = AtoModel(self, self.game.blue.ato)
             self.red_ato_model = AtoModel(self, self.game.red.ato)
+        self.is_ownfor = False
 
         # For UI purposes
         self.allocated_freqs: list[RadioFrequency] = list()