+++ changelog.md
@@ -13,12 +13,12 @@
 * **[Options]** Extend option (so it can be disabled when fixed in DCS) to force air-starts (except for the slots that work) at Ramon Airbase, similar to the Nevatim fix in Retribution 1.3.0
 * **[Options]** New option in Settings: Default start type for Player flights.
 * **[AirWing]** Expose OPFOR Squadrons, giving the ability to change liveries, auto-assignable mission types & an easy way to retrieve debug information.
-
+* **[ATO]** Allow planning as OPFOR
+* 
 ## Fixes
 * **[UI/UX]** A-10A flights can be edited again.
 * **[Mission Generation]** IADS bug sometimes triggering "no skynet usable units" error during mission generation
 * **[New Game Wizard]** Campaign errors show a dialog again and avoid CTDs
-
 
 # Retribution v1.3.1
 #### Note: Re-save your missions in DCS' Mission Editor to avoid possible crashes due to datalink (usually the case when F-16C blk50s are used) when hosting missions on a dedicated server.